requirements.txt
numpy==1.20.3
requests==2.25.1
pydantic==1.7.4
psycopg2-binary==2.9.1
fastapi==0.68.0
fastapi_utils==0.2.1
python-dotenv==0.19.0
requests_oauthlib==1.3.0
SQLAlchemy==1.4.22
torch==1.9.0
transformers==4.9.2
tweepy==3.10.0
uvicorn==0.15.0
